Restructured directories, added .gitignore and CMakeLists.txt. Fixed small bug with default values being declared twice in PlotJob.cpp
