--- compile.sh
@@ -1,1 +1,0 @@
-g++-10 -Wall -Werror -std=c++17 main.cpp Config.cpp jobMan.cpp plotJob.cpp -o job.out
