+++ headers/global_config.hpp
