+++ headers/headers.hpp
@@ -1,0 +1,7 @@
+// Classes
+#include "JobMan.hpp"
+#include "PlotJob.hpp"
+#include "Config.hpp"
+// Other
+#include "global_config.hpp"
+#include "logParser.hpp"
