+++ test/main.cpp
@@ -1,0 +1,25 @@
+
+#include "../headers/headers.hpp"
+
+#include <iostream>
+#include <filesystem>
+
+int main() {
+
+    std::cout << "Test Start" << std::endl;
+
+    //Config Config("config.txt");
+    //Config::configMap localMap = Config.getMappedConfig();
+    //if (!localMap.empty())
+    //    std::cout << "Config passed" << std::endl;
+
+    //std::unordered_map<std::string, std::string> mapbaby = logParser::parsePlotLog("samplePlotOutput.log");
+    //for ( auto [key, val] : mapbaby ){
+    //    std::cout << "Key: " << std::setw(20) << key << " | Val: " << val << std::endl;
+
+        
+    //}
+
+    // Config.writeTest("test", "test.test");
+    //return 0;
+}
